migrate express tests to new autotest api
