--- test/express/fixtures/.gitignore
@@ -1,1 +1,0 @@
-node_modules/
