--- test/express/fixtures/express-4/.gitignore
@@ -1,1 +1,0 @@
-/autotests
