--- test/express/fixtures/express-5/.gitignore
@@ -1,1 +1,0 @@
-/autotests
