--- test/express/fixtures/package.json
@@ -1,7 +1,0 @@
-{
-    "name": "marko-test-express-4",
-    "version": "1.0.0",
-    "devDependencies": {
-        "express": "^4.13.4"
-    }
-}
